Refactor workflows and rename agent health tests to Gm
