--- .github/workflows/TS_AgentHealth.yml
@@ -1,61 +1,0 @@
-name: TS_AgentHealth
-description: "should verify health of agents"
-
-on:
-  schedule:
-    - cron: "0,30 * * * *" # Runs at 0 and 30 minutes past each hour
-  workflow_dispatch:
-  # pull_request:
-  #   branches:
-  #     - main
-
-jobs:
-  TS_AgentHealth:
-    timeout-minutes: 10
-    runs-on: ubuntu-latest
-    env:
-      DATADOG_API_KEY: ${{ secrets.DATADOG_API_KEY }}
-      LOGGING_LEVEL: ${{ vars.LOGGING_LEVEL }}
-      GEOLOCATION: ${{ vars.GEOLOCATION }}
-      GM_BOT_ADDRESS: ${{ vars.GM_BOT_ADDRESS }}
-      WALLET_KEY: ${{ secrets.WALLET_KEY }}
-      ENCRYPTION_KEY: ${{ secrets.ENCRYPTION_KEY }}
-      XMTP_ENV: ${{ vars.XMTP_ENV || 'dev' }}
-      OPENAI_API_KEY: ${{ secrets.OPENAI_API_KEY }}
-      SLACK_BOT_TOKEN: ${{ secrets.SLACK_BOT_TOKEN }}
-      SLACK_CHANNEL: ${{ secrets.SLACK_CHANNEL || 'xmtp-qa' }}
-      TEST_NAME: "TS_AgentHealth"
-    steps:
-      - uses: actions/checkout@v4
-      - name: Debug API key
-        run: |
-          if [ -n "${{ secrets.DATADOG_API_KEY }}" ]; then
-            echo "API key is available in test job"
-          else
-            echo "API key is NOT available in test job"
-          fi
-      - uses: actions/setup-node@v4
-        with:
-          node-version-file: ".node-version"
-          cache: "yarn"
-        env:
-          SKIP_YARN_COREPACK_CHECK: "1"
-      - run: corepack enable
-      - run: yarn
-      - name: Make test script executable
-        run: chmod +x ./scripts/run-test.sh
-      - name: Run tests with retry
-        run: ./scripts/run-test.sh TS_AgentHealth 1
-      - name: Send Slack notification
-        if: always()
-        env:
-          JOB_STATUS: ${{ job.status }}
-        run: npx tsx helpers/slack.ts
-      - name: Upload reports and snapshots artifacts
-        if: always()
-        uses: actions/upload-artifact@v4
-        with:
-          name: ts_agenthealth_${{ matrix.environment }}
-          path: |
-            logs/
-            logs/**/*.png
