+++ .github/workflows/TS_Gm.yml
@@ -1,29 +1,27 @@
 name: TS_Gm
-description: "should verify performance of the library in the dev network"
+description: "should verify health of agents"
 
 on:
   schedule:
-    - cron: "15,45 * * * *" # Runs at 15 and 45 minutes past each hour
+    - cron: "0,30 * * * *" # Runs at 0 and 30 minutes past each hour
   workflow_dispatch:
 
 jobs:
   TS_Gm:
     timeout-minutes: 10
     runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        environment: [production]
     env:
       DATADOG_API_KEY: ${{ secrets.DATADOG_API_KEY }}
       LOGGING_LEVEL: ${{ vars.LOGGING_LEVEL }}
-      XMTP_ENV: ${{ matrix.environment }}
       GEOLOCATION: ${{ vars.GEOLOCATION }}
       GM_BOT_ADDRESS: ${{ vars.GM_BOT_ADDRESS }}
       WALLET_KEY: ${{ secrets.WALLET_KEY }}
       ENCRYPTION_KEY: ${{ secrets.ENCRYPTION_KEY }}
+      XMTP_ENV: ${{ vars.XMTP_ENV || 'dev' }}
       OPENAI_API_KEY: ${{ secrets.OPENAI_API_KEY }}
       SLACK_BOT_TOKEN: ${{ secrets.SLACK_BOT_TOKEN }}
       SLACK_CHANNEL: ${{ secrets.SLACK_CHANNEL || 'xmtp-qa' }}
+      TEST_NAME: "TS_Gm"
     steps:
       - uses: actions/checkout@v4
       - name: Debug API key
@@ -44,7 +42,7 @@ jobs:
       - name: Make test script executable
         run: chmod +x ./scripts/run-test.sh
       - name: Run tests with retry
-        run: ./scripts/run-test.sh TS_Gm
+        run: ./scripts/run-test.sh TS_Gm 1
       - name: Send Slack notification
         if: always()
         env:
@@ -54,7 +52,7 @@ jobs:
         if: always()
         uses: actions/upload-artifact@v4
         with:
-          name: ts_gm_${{ matrix.environment }}
+          name: TS_Gm_${{ matrix.environment }}
           path: |
             logs/
             logs/**/*.png
