+++ README.md
@@ -2,13 +2,12 @@
 
 ## Automated workflows
 
-| Test suite      | Status                                                                                                                                                                            | Resources                                                                                                                                                                       | Run frequency | Networks         |
-| --------------- | --------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- | ------------- | ---------------- |
-| 🚀 Performance  | [![Status](https://github.com/xmtp/xmtp-qa-testing/actions/workflows/TS_Performance.yml/badge.svg)](https://github.com/xmtp/xmtp-qa-testing/actions/workflows/TS_Performance.yml) | [Workflow](https://github.com/xmtp/xmtp-qa-testing/actions/workflows/TS_Performance.yml) / [Test code](https://github.com/xmtp/xmtp-qa-testing/tree/main/suites/TS_Performance) | Every 30 min  | `dev,production` |
-| 📬 Delivery     | [![Status](https://github.com/xmtp/xmtp-qa-testing/actions/workflows/TS_Delivery.yml/badge.svg)](https://github.com/xmtp/xmtp-qa-testing/actions/workflows/TS_Delivery.yml)       | [Workflow](https://github.com/xmtp/xmtp-qa-testing/actions/workflows/TS_Delivery.yml) / [Test code](https://github.com/xmtp/xmtp-qa-testing/tree/main/suites/TS_Delivery)       | Every 30 min  | `dev,production` |
-| 👋 Gm           | [![Status](https://github.com/xmtp/xmtp-qa-testing/actions/workflows/TS_Gm.yml/badge.svg)](https://github.com/xmtp/xmtp-qa-testing/actions/workflows/TS_Gm.yml)                   | [Workflow](https://github.com/xmtp/xmtp-qa-testing/actions/workflows/TS_Gm.yml) / [Test code](https://github.com/xmtp/xmtp-qa-testing/tree/main/suites/TS_Gm)                   | Every 30 min  | `dev,production` |
-| ⏳ Groups       | [![Status](https://github.com/xmtp/xmtp-qa-testing/actions/workflows/TS_Large.yml/badge.svg)](https://github.com/xmtp/xmtp-qa-testing/actions/workflows/TS_Large.yml)             | [Workflow](https://github.com/xmtp/xmtp-qa-testing/actions/workflows/TS_Large.yml) / [Test code](https://github.com/xmtp/xmtp-qa-testing/tree/main/suites/TS_Large)             | Every 2 hours | `production`     |
-| 🚑 Agent health | [![Status](https://github.com/xmtp/xmtp-qa-testing/actions/workflows/TS_AgentHealth.yml/badge.svg)](https://github.com/xmtp/xmtp-qa-testing/actions/workflows/TS_AgentHealth.yml) | [Workflow](https://github.com/xmtp/xmtp-qa-testing/actions/workflows/TS_AgentHealth.yml) / [Test code](https://github.com/xmtp/xmtp-qa-testing/tree/main/suites/TS_AgentHealth) | Every 30 min  | `production`     |
+| Test suite     | Status                                                                                                                                                                            | Resources                                                                                                                                                                       | Run frequency | Networks         |
+| -------------- | --------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- | ------------- | ---------------- |
+| 🚀 Performance | [![Status](https://github.com/xmtp/xmtp-qa-testing/actions/workflows/TS_Performance.yml/badge.svg)](https://github.com/xmtp/xmtp-qa-testing/actions/workflows/TS_Performance.yml) | [Workflow](https://github.com/xmtp/xmtp-qa-testing/actions/workflows/TS_Performance.yml) / [Test code](https://github.com/xmtp/xmtp-qa-testing/tree/main/suites/TS_Performance) | Every 30 min  | `dev,production` |
+| 📬 Delivery    | [![Status](https://github.com/xmtp/xmtp-qa-testing/actions/workflows/TS_Delivery.yml/badge.svg)](https://github.com/xmtp/xmtp-qa-testing/actions/workflows/TS_Delivery.yml)       | [Workflow](https://github.com/xmtp/xmtp-qa-testing/actions/workflows/TS_Delivery.yml) / [Test code](https://github.com/xmtp/xmtp-qa-testing/tree/main/suites/TS_Delivery)       | Every 30 min  | `dev,production` |
+| ⏳ Groups      | [![Status](https://github.com/xmtp/xmtp-qa-testing/actions/workflows/TS_Large.yml/badge.svg)](https://github.com/xmtp/xmtp-qa-testing/actions/workflows/TS_Large.yml)             | [Workflow](https://github.com/xmtp/xmtp-qa-testing/actions/workflows/TS_Large.yml) / [Test code](https://github.com/xmtp/xmtp-qa-testing/tree/main/suites/TS_Large)             | Every 2 hours | `production`     |
+| 🤖 Gm          | [![Status](https://github.com/xmtp/xmtp-qa-testing/actions/workflows/TS_Gm.yml/badge.svg)](https://github.com/xmtp/xmtp-qa-testing/actions/workflows/TS_Gm.yml)                   | [Workflow](https://github.com/xmtp/xmtp-qa-testing/actions/workflows/TS_Gm.yml) / [Test code](https://github.com/xmtp/xmtp-qa-testing/tree/main/suites/TS_Gm)                   | Every 30 min  | `production`     |
 
 ## Testing scope
