+++ functional/browser.test.ts
@@ -1,4 +1,5 @@
 import { loadEnv } from "@helpers/client";
+import generatedInboxes from "@helpers/generated-inboxes.json";
 import { logError } from "@helpers/logger";
 import { XmtpPlaywright } from "@helpers/playwright";
 import { beforeAll, describe, expect, it } from "vitest";
@@ -38,8 +39,11 @@ describe(testName, () => {
       if (!gmBotAddress) {
         throw new Error("GM_BOT_ADDRESS environment variable is not set");
       }
-      await xmtpTester.newGroupFromUI(gmBotAddress);
-      await xmtpTester.fillAddressesAndCreate([gmBotAddress]);
+      const slicedInboxes = generatedInboxes.slice(0, 4);
+      await xmtpTester.newGroupFromUI([
+        ...slicedInboxes.map((inbox) => inbox.accountAddress),
+        gmBotAddress,
+      ]);
       await xmtpTester.sendMessage("hi");
       await xmtpTester.waitForResponse("gm");
     } catch (e) {
